Revert "Merge branch 'bug-fixes' into feature/view-recommended-screenings"
This reverts commit 0568278da2c2109490300a62d2214675e1a65490, reversing
changes made to bec93f05f2f068375242c3048cff46852680465a.
